Move ReportHPXMLOutput into UpgradeCosts.
