--- measures/HPXMLOutput/README.md
@@ -1,30 +1,0 @@
-
-###### (Automatically generated documentation)
-
-# HPXML Output
-
-## Description
-Reports HPXML outputs for residential HPXML-based models.
-
-Parses the HPXML file and reports pre-defined outputs.
-
-## Arguments
-
-
-**Output Format**
-
-The file format of the annual (and timeseries, if requested) outputs.
-
-- **Name:** ``output_format``
-- **Type:** ``Choice``
-
-- **Required:** ``false``
-
-- **Choices:** `csv`, `json`, `msgpack`
-
-<br/>
-
-
-
-
-
