--- measures/HPXMLOutput/measure.xml
@@ -1,91 +1,0 @@
-<?xml version="1.0"?>
-<measure>
-  <schema_version>3.1</schema_version>
-  <name>hpxml_output</name>
-  <uid>9561a0d7-60ad-48c5-8337-2461df044d80</uid>
-  <version_id>044aa2c7-7196-4ee5-850c-fba604a915fd</version_id>
-  <version_modified>2024-06-06T20:38:31Z</version_modified>
-  <xml_checksum>B1F14CB4</xml_checksum>
-  <class_name>HPXMLOutput</class_name>
-  <display_name>HPXML Output</display_name>
-  <description>Reports HPXML outputs for residential HPXML-based models.</description>
-  <modeler_description>Parses the HPXML file and reports pre-defined outputs.</modeler_description>
-  <arguments>
-    <argument>
-      <name>output_format</name>
-      <display_name>Output Format</display_name>
-      <description>The file format of the annual (and timeseries, if requested) outputs.</description>
-      <type>Choice</type>
-      <required>false</required>
-      <model_dependent>false</model_dependent>
-      <default_value>csv</default_value>
-      <choices>
-        <choice>
-          <value>csv</value>
-          <display_name>csv</display_name>
-        </choice>
-        <choice>
-          <value>json</value>
-          <display_name>json</display_name>
-        </choice>
-        <choice>
-          <value>msgpack</value>
-          <display_name>msgpack</display_name>
-        </choice>
-      </choices>
-    </argument>
-  </arguments>
-  <outputs />
-  <provenances />
-  <tags>
-    <tag>Reporting.QAQC</tag>
-  </tags>
-  <attributes>
-    <attribute>
-      <name>Measure Type</name>
-      <value>ModelMeasure</value>
-      <datatype>string</datatype>
-    </attribute>
-    <attribute>
-      <name>Intended Software Tool</name>
-      <value>OpenStudio Application</value>
-      <datatype>string</datatype>
-    </attribute>
-    <attribute>
-      <name>Intended Software Tool</name>
-      <value>Parametric Analysis Tool</value>
-      <datatype>string</datatype>
-    </attribute>
-  </attributes>
-  <files>
-    <file>
-      <filename>README.md</filename>
-      <filetype>md</filetype>
-      <usage_type>readme</usage_type>
-      <checksum>1AE3E7BD</checksum>
-    </file>
-    <file>
-      <filename>README.md.erb</filename>
-      <filetype>erb</filetype>
-      <usage_type>readmeerb</usage_type>
-      <checksum>513F28E9</checksum>
-    </file>
-    <file>
-      <version>
-        <software_program>OpenStudio</software_program>
-        <identifier>3.2.0</identifier>
-        <min_compatible>3.2.0</min_compatible>
-      </version>
-      <filename>measure.rb</filename>
-      <filetype>rb</filetype>
-      <usage_type>script</usage_type>
-      <checksum>3F8895AA</checksum>
-    </file>
-    <file>
-      <filename>constants.rb</filename>
-      <filetype>rb</filetype>
-      <usage_type>resource</usage_type>
-      <checksum>64289999</checksum>
-    </file>
-  </files>
-</measure>
