+++ measures/ResStockArguments/measure.xml
@@ -3,8 +3,8 @@
   <schema_version>3.1</schema_version>
   <name>res_stock_arguments</name>
   <uid>c984bb9e-4ac4-4930-a399-9d23f8f6936a</uid>
-  <version_id>d5267b62-b346-4d24-8fee-32ba8bb30203</version_id>
-  <version_modified>2024-06-07T14:50:13Z</version_modified>
+  <version_id>f334dcd7-0413-4229-930e-7adcb410aa89</version_id>
+  <version_modified>2024-06-07T20:21:15Z</version_modified>
   <xml_checksum>2C38F48B</xml_checksum>
   <class_name>ResStockArguments</class_name>
   <display_name>ResStock Arguments</display_name>
@@ -7503,7 +7503,7 @@
       <filename>constants.rb</filename>
       <filetype>rb</filetype>
       <usage_type>resource</usage_type>
-      <checksum>9740F988</checksum>
+      <checksum>AB1AADDC</checksum>
     </file>
     <file>
       <filename>resstock_arguments_test.rb</filename>
