+++ measures/ResStockArguments/resources/constants.rb
@@ -97,6 +97,7 @@ class Constants
   end
 
   def self.arguments_to_register
+    # list of resstock arguments; reported as build_existing_model.<argument_name>, ...
     return []
   end
 end
