+++ measures/UpgradeCosts/measure.xml
@@ -3,8 +3,8 @@
   <schema_version>3.1</schema_version>
   <name>upgrade_costs</name>
   <uid>ef51212c-acc4-48d7-9b29-cf2a5c6c4449</uid>
-  <version_id>1542f7b0-79d1-4346-8835-c0dc48393f2d</version_id>
-  <version_modified>2024-06-06T21:55:02Z</version_modified>
+  <version_id>c8b84aa9-5d68-40ff-88e4-495eab22ed69</version_id>
+  <version_modified>2024-06-07T20:41:25Z</version_modified>
   <xml_checksum>B1F14CB4</xml_checksum>
   <class_name>UpgradeCosts</class_name>
   <display_name>Upgrade Costs</display_name>
@@ -75,7 +75,13 @@
       <filename>measure.rb</filename>
       <filetype>rb</filetype>
       <usage_type>script</usage_type>
-      <checksum>B51C9EF9</checksum>
+      <checksum>B50B2AE9</checksum>
+    </file>
+    <file>
+      <filename>constants.rb</filename>
+      <filetype>rb</filetype>
+      <usage_type>resource</usage_type>
+      <checksum>64289999</checksum>
     </file>
     <file>
       <filename>MF_1story_UB_Furnace_AC1_FuelTankWH.osw</filename>
@@ -210,10 +216,22 @@
       <checksum>9494BFFE</checksum>
     </file>
     <file>
+      <filename>results_annual.csv</filename>
+      <filetype>csv</filetype>
+      <usage_type>test</usage_type>
+      <checksum>CD046FB5</checksum>
+    </file>
+    <file>
+      <filename>results_design_load_details.csv</filename>
+      <filetype>csv</filetype>
+      <usage_type>test</usage_type>
+      <checksum>CDB8FD7B</checksum>
+    </file>
+    <file>
       <filename>upgrade_costs_test.rb</filename>
       <filetype>rb</filetype>
       <usage_type>test</usage_type>
-      <checksum>AD8DE741</checksum>
+      <checksum>5087DD5F</checksum>
     </file>
   </files>
 </measure>
