+++ workflow/run_analysis.rb
@@ -287,10 +287,6 @@ def run_workflow(yml, in_threads, measures_only, debug_arg, overwrite, building_
         }
       },
       {
-        'measure_dir_name' => 'HPXMLOutput',
-        'arguments' => { 'output_format' => 'csv' }
-      },
-      {
         'measure_dir_name' => 'UpgradeCosts',
         'arguments' => { 'debug' => debug }
       }
